requirements.txt
appnope==0.1.2
backcall==0.2.0
decorator==4.4.2
Django==2.2
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.0.0
django-redis==4.3.0
djangorestframework==3.9.4
gunicorn==20.0.4
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pytz==2021.1
redis==2.10.3
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
psycopg2-binary==2.8.6
requests==2.25.1